# Specialize `setindex` and `_totuple` for isbits `NTuple`s