--- a/base/tuple.jl
+++ b/base/tuple.jl
@@ -56,13 +56,12 @@ true
 function setindex(x::T, v::V, i::Integer) where {T<:Tuple, V}
     @boundscheck 1 <= i <= length(x) || throw(BoundsError(x, i))
     @inline
-    T1 = @inbounds fieldtype(T, Int(i))
-    if V <: T1 && allocatedinline(T1)
+    if @inbounds fieldtype(T, Int(i)) == V && isbitstype(V)
         new_tuple = Ref(x)
-        t = @_gc_preserve_begin new_tuple
         new_tuple_ptr = unsafe_convert(Ptr{T}, new_tuple)
-        ith_index_ptr = Ptr{T1}(new_tuple_ptr + fieldoffset(T, i))
-        unsafe_store!(ith_index_ptr, v, 1)
+        ith_index_ptr = Ptr{V}(new_tuple_ptr + fieldoffset(T, i))
+        t = @_gc_preserve_begin new_tuple
+        unsafe_store!(ith_index_ptr, v)
         @_gc_preserve_end t
         new_tuple[]
     else
@@ -398,9 +397,9 @@ function _totuple_err(@nospecialize T)
     throw(ArgumentError("too few elements for tuple type $T"))
 end
 
-function _tontuple_inline(::Type{NTuple{N, T}}, itr) where {N, T}
+function _tontuple_isbits(::Type{NTuple{N, T}}, itr) where {N, T}
     @inline
-    allocatedinline(T) || throw(ArgumentError("tuple type $T is not allocated inline"))
+    isbitstype(T) || throw(ArgumentError("tuple type $T is not isbits"))
     tuple = Ref{NTuple{N, T}}()
     tuple_ptr = Ptr{T}(unsafe_convert(Ptr{NTuple{N, T}}, tuple))
 
@@ -419,8 +418,8 @@ end
 
 function _totuple(::Type{NTuple{N, T}}, itr) where {N, T}
     @inline
-    if allocatedinline(T)
-        _tontuple_inline(NTuple{N, T}, itr)
+    if isbitstype(T)
+        _tontuple_isbits(NTuple{N, T}, itr)
     elseif N >= 32
         # use iterative algorithm for long tuples
         elts = collect(E, Iterators.take(itr,N))