This pr instead of using a recursive solution like the existing method, creates a `Ref` to a tuple and mutates that iteratively. Escape Analysis seems to work well enough that this is always stack allocated.

master
```julia
julia>  @btime Base.setindex($(Ref{NTuple{100, Int}}()[];), 0, 5);  
  1.520 μs (98 allocations: 3.09 KiB)

julia>  @btime Base.setindex($(Ref{NTuple{5, Int}}()[];), 0, 5);  
  2.100 ns (0 allocations: 0 bytes)

julia> @btime NTuple{100, Int}($(i for i in 1:100));
  2.067 μs (2 allocations: 1.67 KiB)

julia> @btime NTuple{5, Int}($(i for i in 1:5));
  2.300 ns (0 allocations: 0 bytes)

julia> @btime NTuple{28, Int}($(i for i in 1:28));
  13.614 ns (0 allocations: 0 bytes)

julia> @btime NTuple{28, Float64}((rand()*i for i in 1:28));
  74.409 ns (0 allocations: 0 bytes)
```
pr
```julia
julia> @btime Base.setindex($(Ref{NTuple{100, Int}}()[];), 0, 5);
  13.313 ns (0 allocations: 0 bytes)

julia> @btime Base.setindex($(Ref{NTuple{5, Int}}()[];), 0, 5);
  2.000 ns (0 allocations: 0 bytes)

julia> @btime NTuple{100, Int}($(i for i in 1:100));
  24.574 ns (0 allocations: 0 bytes)

julia> @btime NTuple{5, Int}($(i for i in 1:5));
  2.500 ns (0 allocations: 0 bytes)

julia> @btime NTuple{28, Int}($(i for i in 1:28));
  8.709 ns (0 allocations: 0 bytes)

julia> @btime NTuple{28, Float64}((rand()*i for i in 1:28));
  45.510 ns (0 allocations: 0 bytes)
```

The new code seems to be about 100x faster for sizes >= 32 and it also SIMDs better. We could also specialize `_totuple` for isbits `Tuples` but that requires unrolling to prevent type instabilities and runtime dispatch.