@Zentrik 2023-10-17
Not really sure why this test failed 
```jl
ambiguous                                        (4) \|         failed at 2023-10-17T15:36:44.229
--
  | Test Failed at /Users/julia/.julia/scratchspaces/a66863c6-20e8-4ff4-8a62-49f30b1f605e/agent-cache/default-honeycrisp-XC9YQX9HH2.0/build/default-honeycrisp-XC9YQX9HH2-0/julialang/julia-master/julia-7026194d4d/share/julia/test/ambiguous.jl:360
  | Expression: need_to_handle_undef_sparam == Set()
  | Evaluated: Set(Method[_tontuple_isbits(::Type{NTuple{N, T}}, itr) where {N, T} @ Base tuple.jl:400, _totuple(::Type{NTuple{N, T}}, itr) where {N, T} @ Base tuple.jl:419]) == Set{Any}()
```
I'm pretty sure I'm using `N` and `T` in both functions, is this just a bug in the test?