@Zentrik 2023-10-20
I have a couple questions, I want to add support for non isbitstype and I think the main change is adding a GC write barrier, can this be done in Julia/ as an intrinsic or do I need to write that part in C? Also are intrinsics optimized by the compiler as well as equivalent code in Julia?